#31 code-review fixes.
- Move %doit code & docs from `doit_cmd.py`-->`tools.py`.
- doc: Undo mistaken replacements in pre-previous commit. 
- doc: Add IPython-integration into "features", link with IPython docs,
update CHANGES.
- doc: Add me(ankostis) into AUTHORS.
- spell: Update dictionary.
