+++ AUTHORS
@@ -17,3 +17,4 @@
  * Chris Warrick - Kwpolska - http://chriswarrick.com/
  * Ronan Le Gallic - rolegic <at> gmail <dot> com
  * Simon Conseil - contact <at> saimon <dot> org
+ * Kostis Anagnostopoulos - ankostis <at> gmail <dot> com
