+++ CHANGES
@@ -8,6 +8,7 @@ Changes
 
  - fix GH-#22: Allow to customize how file_dep are checked
  - BACKWARD INCOMPATIBLE: signature for custom DB backend changed
+ - GH-#31: Add IPython `%doit` magic-function loading tasks from its global namespace.
 
 
 0.27.0 (*2015-01-30*)
