+++ doit/doit_cmd.py
@@ -132,53 +132,3 @@ class DoitMain(object):
                 pdb.post_mortem(sys.exc_info()[2])
             sys.stderr.write(traceback.format_exc())
             return 3
-
-
-def register_doit_as_IPython_magic():
-    """
-    Defines a ``%doit`` magic function that discovers and execute tasks  
-    from IPython's interactive variables (global namespace).
-    
-    It will fail if not invoked from within an interactive IPython shell.
-    
-    .. Tip::
-        To permanently add this magic-function to your IPython, create a new script 
-        inside your startup-profile (``~/.ipython/profile_default/startup/doit_magic.ipy``) 
-        with the following content:
-        
-            from doit.doit_cmd import register_doit_as_IPython_magic
-            register_doit_as_IPython_magic()
-
-    """
-    from IPython.core.magic import register_line_magic
-
-    from doit.cmd_base import ModuleTaskLoader
-    
-    @register_line_magic
-    def doit(line):
-        """
-        Run *doit* with `task_creators` from all interactive variables (IPython's global namespace).
-        
-        Examples:
-        
-            >>> %doit --help          ## Show help for options and arguments.
-
-            >>> def task_foo():
-                    return {'actions': ['echo hi IPython'], 
-                            'verbosity': 2}
-            
-            >>> %doit list            ## List any tasks discovered.
-            foo
-            
-            >>> %doit                 ## Run any tasks.
-            .  foo
-            hi IPython
-
-        """
-        try:
-            import get_ipython ## Workaround pyflake complaints.
-        except ImportError:
-            pass
-        ip = get_ipython()
-        commander       = DoitMain(ModuleTaskLoader(ip.user_module))
-        commander.run(line.split())
